+++ tests/dummy/app/routes/components-acceptance-tests/flexberry-objectlistview/base-operations.js
@@ -3,6 +3,7 @@ import ListFormRoute from 'ember-flexberry/routes/list-form';
 export default ListFormRoute.extend({
   /**
     Name of model projection to be used as record's properties limitation.
+
     @property modelProjection
     @type String
     @default 'SuggestionL'
@@ -24,6 +25,7 @@ export default ListFormRoute.extend({
   }
   For default userSetting use empty name ('').
   <componentName> may contain any of properties: colsOrder, sorting, colsWidth or being empty.
+
   @property developerUserSettings
   @type Object
   @default {}
@@ -32,6 +34,7 @@ export default ListFormRoute.extend({
 
   /**
     Name of model to be used as list's records types.
+    
     @property modelName
     @type String
     @default 'ember-flexberry-dummy-suggestion'